#291 Updates to user and activity models
